Added replay, fixed user input errors
Added replay option for users to decide if they want to replay the
game. Fixed the user input prompting for pickSymbols() so that it
reprints the prompt in case of bad user input. Added a catch for
ArrayIndexOutOfBounds exception in the event that a user inputs a
number less than 0 or more than 8. Issue with “play again” printing
twice, have to figure that out.